Major refactor and clean up of functions in bot.py . Moved/Grouped similar functions into seperate files within utils folder.
